Updated the main controller
Added a single function which will be called and passed a buttons id to know which function to call while simultaneously calling functions to toggle off buttons, toggle the sorting text, toggle the clock to start running and an interval to update the time. When the promise is resolved the interval is then cleared, toggle the text to sorted and then reset buttons. We then do a quick loop to change the colors of the array and show it is sorted.